Fix serial generating and parsing
Jarolift serials are 28bit values.
Currently, the serials are being generated from a 24bit input that is shifted by 8 bits. This causes the following problems:

The 4 highest bits of the input are being ignored. Inputs 0x00BC61 and 0xF0BC61 will effectively send the same serials to the covers (although the values stored in the EEPROM will differ).
It is impossible to 'copy' a remote if its serials do not end in 0x, x being the increasing value 0 through F. e.g. A remote with channel 0 serial 0x0BC6110 can not be copied.
This PR corrects both problems by shifting the input value by 4 bits instead of 8. In the cmd_ functions, we now use the actual serial from the EEPROM instead of the serials[] array to calculate the last byte used in the encrypted part of the final message as it has to include the last 4 bits of the input value.

The current serials are stored in the EEPROM in full, so this PR should not break current configurations until serials are regenerated with the same input. If serials must be regenerated after this patch, shift the input and add a 0: 0x00BC61 becomes 0x0BC610. The webui should reflect this change automatically.